move Simulation to controller package
