+++ src/main/java/com/codecool/App.java
@@ -1,10 +1,7 @@
 package com.codecool;
 
 
-import com.codecool.model.Simulation;
-import com.codecool.model.board.Board;
-import com.codecool.view.View;
-import com.codecool.view.windowedview.WindowedView;
+import com.codecool.controller.Simulation;
 
 public class App {
